% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3dR.R
\name{v3dR}
\alias{v3dR}
\title{v3dR}
\usage{
v3dR(full_filepath)
}
\arguments{
\item{full_filepath}{Full filepath to txt file}
}
\value{
Dataframe containing txt data in long format
}
\description{
Import V3D ASCII Txt File To R
}
\examples{
full_filepath = file.path("C:/local_repos/github/v3dR/data/L_A.txt")
df <- v3dR(full_filepath)
}
